requirements-dev.txt
-r requirements.txt
django-filter~=23.3
parameterized==0.9.0
coverage==7.3.2